Split runtime review parity gates
